Check not are on a docker container.

+++ stopDevelopmentEnvironment.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-pushd $DIR >/dev/null
-docker stop wenet_common_dev
-docker rm wenet_common_dev
-popd >/dev/null
+if [ -f /.dockerenv ]; then
+   echo "You can not stop the development environment inside a docker container"
+else
+	DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+	pushd $DIR >/dev/null
+	docker stop wenet_common_dev
+	docker rm wenet_common_dev
+	popd >/dev/null
+fi
